+++ src/video_core/renderer_vulkan/host_passes/fsr_pass.h
@@ -15,7 +15,7 @@ public:
     struct Settings {
         bool enable{true};
         bool use_rcas{true};
-        float rcas_attenuation{Config::getRcasAttenuation()};
+        int rcas_attenuation{Config::getRcasAttenuation()};
     };
 
     void Create(vk::Device device, VmaAllocator allocator, u32 num_images);
